rename file that defines writeObjectTableTask
